refactor: change no ssr components file location
